feat: restructure components and add landing page, navbar, and popups with styling
