+++ src/components/SignupPopup/SignupPopup.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "./SignupPopup.css";
-import THLogo from "../assets/TH-Scribes-Logo.png"; // Import the logo
+import THLogo from "../../assets/TH-Scribes-Logo.png"; // Import the logo
 
 interface SignupPopupProps {
   onClose: () => void;
